requirements.txt
chainlit>=1.3.0
python-engineio==4.9.1
python-socketio==5.11.0
google-genai
pandas>=2.2.0
python-dotenv>=1.0.1
anthropic>=0.50.0
yfinance>=0.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
fastapi
uvicorn[standard]
pytest>=9.0.0
